% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-items.R
\name{decryptItem}
\alias{decryptItem}
\title{Decrypt a credential item}
\usage{
decryptItem(item)
}
\arguments{
\item{item}{A credential item to be decrypted.}
}
\value{
A decrypted credential item.
}
\description{
Due to insecurity of file based credential storage, the
             sensitive information is encrypted.
}

