% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based-storage.R
\name{encryptItem}
\alias{encryptItem}
\title{Encrypt a credential item}
\usage{
encryptItem(item)
}
\arguments{
\item{item}{A credential item to be encrypted.}
}
\value{
An encrypted credential item.
}
\description{
Due to insecurity of file based credential storage, the
             sensitive information is encrypted.
}

