% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based-storage.R
\name{fetchFileBased}
\alias{fetchFileBased}
\title{Fetch a credential item from file-based storage}
\usage{
fetchFileBased(id)
}
\arguments{
\item{id}{The ID of the requested credential item.}
}
\value{
The requested credential item.
}
\description{
Fetches a credential item from secure file-based storage.
}

