% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileBasedStorageSession}
\alias{fileBasedStorageSession}
\title{Make file-based storage available}
\usage{
fileBasedStorageSession()
}
\value{
NULL (invisibly).
}
\description{
Ensures that all necessary infomration for file-based storage
             is available (secretr.path and secretr.key). If the
             .Rsecrets file path is not available, file-based storage is
             initialized and if either the key or the iv vector are not
             available, the user is asked for this information.
}
