% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generatePassword}
\alias{generatePassword}
\title{Generate a random password}
\usage{
generatePassword(length = 15, lowercase = TRUE, uppercase = TRUE,
  digits = TRUE, symbols = TRUE)
}
\arguments{
\item{length}{The length of the resulting password.}

\item{lowercase}{Whether to include lowercase characters (logical).}

\item{uppercase}{Whether to include uppercase characters (logical).}

\item{digits}{Whether to include digits (logical).}

\item{symbols}{Whether to include special characters (logical).}
}
\value{
A password as a string.
}
\description{
A password is generated by sampling letters from an alphabet,
             consisting any combination of lowercase and uppercase
             characters, digits as well as special characters.
}

