% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-items.R
\name{getItemId}
\alias{getItemId}
\title{Get the ID of an Item}
\usage{
getItemId(item)
}
\arguments{
\item{item}{A credential item to be identified.}
}
\value{
The ID of the supplied item.
}
\description{
The ID of each Item is used as a unique identifier.
}

