% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSetItem}
\alias{getSetItem}
\title{Item getter/setter depending on item existence}
\usage{
getSetItem(service_name, credential_type = "LoginItem")
}
\value{
A credential item, such as a LoginItem.
}
\description{
Depending on whether the requested credential item is available
             it is either fetched or created and returned.
}

