% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initFileBasedStorage}
\alias{initFileBasedStorage}
\title{Initialize file-based storage}
\usage{
initFileBasedStorage(key = NULL, secretsPath = file.path("~", ".Rsecrets"),
  profilePath = file.path("~", ".Rprofile"))
}
\arguments{
\item{key}{Key used for en-/decryption. If not user-supplied (NULL),
a new one is generated.}

\item{secretsPath}{Path to the .Rsecrets file (default is in user home).}

\item{profilePath}{Path to the .Rprofile file (default is in user home).}
}
\value{
NULL (invisibly).
}
\description{
In case no OS resource for storing credentials is available/
             implemented, a file-based fall-back is used. Sensitive
             information is encrypted prior to being stored to an .Rsecrets
             file. The key for decryption is either managed by the user
             or written to the user's .Rprofile.
}

