% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential-items.R
\name{LoginItem}
\alias{LoginItem}
\title{Create a LoginItem object}
\usage{
LoginItem(service, username, password = NULL)
}
\arguments{
\item{service}{The name, by which the LoginItem object is identified.}

\item{username}{A string, specifying the user name.}

\item{password}{NULL or a string specifying the password.}
}
\value{
A LoginItem object, which is a list containing entries
        service, username, password and encrypted (logical).
}
\description{
A LoginItem object is created from a service name (which
             is used to identify the credential item), a user name and
             (optionally) a password. If the password is not supplied, a
             new one is generated using generatePassword.
}

