% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readUserInput}
\alias{readUserInput}
\title{A wrapper around readline that supports masking}
\usage{
readUserInput(secret = FALSE, prefix = "")
}
\value{
The string entered by the user.
}
\description{
In case the user enters sensitive information such as a
             password, this function tries to mask the input by calling
             getPass of the eponymous package if it can be made available.
}
