% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based-storage.R
\name{removeFileBased}
\alias{removeFileBased}
\title{Delete a credential item from file-based storage}
\usage{
removeFileBased(id)
}
\arguments{
\item{id}{The ID of the credential item to be deleted.}
}
\value{
NULL (invisibly).
}
\description{
Remove a credential item from secure file-based storage.
}

