% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-item.R
\name{removeItem}
\alias{removeItem}
\title{Delete a credential item}
\usage{
removeItem(id)
}
\arguments{
\item{id}{The ID of the credential item to be deleted.}
}
\value{
NULL (invisibly).
}
\description{
Remove a credential item from secure storage.
}

