% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based-storage.R
\name{storeFileBased}
\alias{storeFileBased}
\title{Store a credential item in file-based storage}
\usage{
storeFileBased(item)
}
\arguments{
\item{item}{A credential item to be stored.}
}
\value{
NULL (invisibly).
}
\description{
Write a credential item to file-based storage, in case no OS
             resource for storing credentials is available/implemented.
}

