% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store-item.R
\name{storeItem}
\alias{storeItem}
\title{Store a credential item}
\usage{
storeItem(item)
}
\arguments{
\item{item}{A credential item to be stored.}
}
\value{
NULL (invisibly).
}
\description{
In order to store credential items, OS resources (such as the
             Mac OS Keychain) are used when possible. As a fall-back option,
             a file-based storage option is available which encrypts
             sensitive information and saves the key to the user's
             .Rprofile.
}

