% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-based-storage.R
\name{updateFileBased}
\alias{updateFileBased}
\title{Update a credential item in file-based storage}
\usage{
updateFileBased(item)
}
\arguments{
\item{item}{A credential item to be modified.}
}
\value{
NULL (invisibly).
}
\description{
In case an already existing credential item changes, this
             function may be used to modify its file-based storage
             representation.
}

