% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-item.R
\name{updateItem}
\alias{updateItem}
\title{Update a credential item}
\usage{
updateItem(item)
}
\arguments{
\item{item}{A credential item to be modified.}
}
\value{
NULL (invisibly).
}
\description{
In case an already existing credential item changes, this
             function may be used to modify its stored representation.
}

