% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-item.R
\name{fetchItem}
\alias{fetchItem}
\title{Fetch a credential item}
\usage{
fetchItem(id, vault = defaultStorageMode())
}
\arguments{
\item{id}{The ID of the requested credential item.}

\item{vault}{A string indicating where the requested credential is stored.}
}
\value{
The requested credential item.
}
\description{
Fetches a credential item from secure storage.
}

